requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via langchain
atlassian-python-api==3.41.21
    # via cc-page-rater (pyproject.toml)
azure-ai-inference==1.0.0b9
    # via cc-page-rater (pyproject.toml)
azure-core==1.32.0
    # via azure-ai-inference
beautifulsoup4==4.13.3
    # via
    #   cc-page-rater (pyproject.toml)
    #   atlassian-python-api
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
cli-exit-tools==1.2.7
    # via wrapt-timeout-decorator
click==8.1.8
    # via
    #   cli-exit-tools
    #   rich-toolkit
    #   typer
    #   uvicorn
deprecated==1.2.18
    # via atlassian-python-api
dill==0.3.9
    # via
    #   multiprocess
    #   wrapt-timeout-decorator
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.11
    # via cc-page-rater (pyproject.toml)
fastapi-cli==0.0.7
    # via fastapi
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   langsmith
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
isodate==0.7.2
    # via azure-ai-inference
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.6
    # via
    #   cc-page-rater (pyproject.toml)
    #   fastapi
jmespath==1.0.1
    # via atlassian-python-api
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.21
    # via cc-page-rater (pyproject.toml)
langchain-core==0.3.47
    # via
    #   langchain
    #   langchain-text-splitters
langchain-text-splitters==0.3.7
    # via langchain
langsmith==0.3.18
    # via
    #   langchain
    #   langchain-core
lib-detect-testenv==2.0.8
    # via
    #   cli-exit-tools
    #   wrapt-timeout-decorator
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multiprocess==0.70.17
    # via wrapt-timeout-decorator
oauthlib==3.2.2
    # via
    #   atlassian-python-api
    #   requests-oauthlib
orjson==3.10.15
    # via
    #   fastapi
    #   langsmith
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
psutil==7.0.0
    # via wrapt-timeout-decorator
pydantic==2.10.6
    # via
    #   cc-page-rater (pyproject.toml)
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-extra-types==2.10.3
    # via fastapi
pydantic-settings==2.8.1
    # via fastapi
pygments==2.19.1
    # via rich
python-dotenv==1.0.1
    # via
    #   cc-page-rater (pyproject.toml)
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via
    #   fastapi
    #   langchain
    #   langchain-core
    #   uvicorn
requests==2.32.3
    # via
    #   cc-page-rater (pyproject.toml)
    #   atlassian-python-api
    #   azure-core
    #   langchain
    #   langsmith
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==2.0.0
    # via atlassian-python-api
requests-toolbelt==1.0.0
    # via langsmith
rich==13.9.4
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.13.2
    # via fastapi-cli
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   atlassian-python-api
    #   azure-core
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.39
    # via langchain
starlette==0.46.1
    # via fastapi
tenacity==9.0.0
    # via langchain-core
toml==0.10.2
    # via cc-page-rater (pyproject.toml)
typer==0.15.2
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   anyio
    #   azure-ai-inference
    #   azure-core
    #   beautifulsoup4
    #   fastapi
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich
    #   rich-toolkit
    #   sqlalchemy
    #   starlette
    #   typer
    #   uvicorn
ujson==5.10.0
    # via fastapi
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via
    #   cc-page-rater (pyproject.toml)
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wrapt==1.17.2
    # via
    #   deprecated
    #   wrapt-timeout-decorator
wrapt-timeout-decorator==1.5.1
    # via cc-page-rater (pyproject.toml)
zstandard==0.23.0
    # via langsmith
